+++ odoo/local-src/sale_cancel_remaining/models/__init__.py
@@ -2,3 +2,4 @@
 # Copyright 2018 Okia SPRL
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 from . import sale_order_line
+from . import stock_move