+++ odoo/local-src/sale_cancel_remaining/models/stock_move.py
@@ -1,0 +1,20 @@
+# -*- coding: utf-8 -*-
+# Copyright 2019 Camptocamp S.A.
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+
+from odoo import models, api
+
+
+class StockMove(models.Model):
+    _inherit = 'stock.move'
+
+    @api.multi
+    def action_done(self):
+        result = super(StockMove, self).action_done()
+        for move in self.filtered('origin_returned_move_id'):
+            line = move.procurement_id.sale_line_id
+            if move.product_id.expense_policy != 'no' or not line:
+                continue
+            line.product_qty_returned += move.product_uom_qty
+        return result