+++ odoo/songs/upgrade/10_30_23_update.py
@@ -15,5 +15,17 @@ def reload_translation(ctx):
 
 
 @anthem.log
+def record_returned_qty(ctx):
+    """populate the returned qty field on SO lines"""
+    returns = ctx.env['stock.move'].search(
+        [('origin_returned_move_id', '!=', False),
+         ('state', '=', 'done'),
+         ]
+    )
+    for move in returns.filtered('order_line_id'):
+        move.order_line_id.product_qty_returned += move.product_uom_qty
+
+
+@anthem.log
 def post(ctx):
     reload_translation(ctx)